@@ -2,7 +2,7 @@ plugins {
     id "idea"
     id "maven-publish"
     id "io.spring.dependency-management" version "1.0.9.RELEASE" apply false
-    id "io.freefair.lombok" version "5.3.0" apply false
+    id "io.freefair.lombok" version "6.0.0-m2" apply false
     id "com.github.johnrengelman.shadow" version "6.1.0" apply false
     id "com.palantir.docker" version "0.25.0" apply false
     id "com.google.protobuf" version "0.8.14" apply false
@@ -45,8 +45,8 @@ subprojects {
     }
 
     java {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
+        sourceCompatibility = JavaVersion.VERSION_16
+        targetCompatibility = JavaVersion.VERSION_16
         withJavadocJar()
         withSourcesJar()
     }
